🔥(backend) remove backend login pages
Remove the backend pages which have been rewrite as
API endpoints for the standalone site.
