--- src/backend/marsha/account/templates/account/password_reset_done.html
@@ -1,11 +1,0 @@
-{% extends 'account/base.html' %}
-{% load i18n %}
-
-{% block content %}
-
-<div id="content-main" class="inner-container">
-  <p>{% translate 'We’ve emailed you instructions for setting your password, if an account exists with the email you entered. You should receive them shortly.' %}</p>
-  <p>{% translate 'If you don’t receive an email, please make sure you’ve entered the address you registered with, and check your spam folder.' %}</p>
-</div>
-
-{% endblock %}
