--- src/backend/marsha/account/templates/account/password_reset_form.html
@@ -1,20 +1,0 @@
-{% extends 'account/base.html' %}
-{% load i18n static %}
-
-{% block content %}
-
-<div id="content-main" class="inner-container">
-  <p>{% translate 'Forgotten your password? Enter your email address below, and we’ll email instructions for setting a new one.' %}</p>
-  <form method="post">{% csrf_token %}
-    <div class="form-row">
-      {{ form.email.errors }}
-      <label for="id_email">{% translate 'Email address:' %}</label>
-      {{ form.email }}
-    </div>
-    <div class="submit-row">
-      <button class="btn-primary" type="submit" value="{% translate 'Reset my password' %}">{% translate 'Reset my password' %}</button>
-    </div>
-  </form>
-</div>
-
-{% endblock %}
